Added DropFilter as a separate regularizer
